pyproject.toml
[project]
name = "auto-certy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi>=0.116.1",
    "fpdf2>=2.8.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pdf2image>=1.17.0",
    "pdfrw>=0.4",
    "pillow>=11.3.0",
    "pydantic[email]>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "reportlab>=4.4.3",
    "requests>=2.32.4",
    "streamlit>=1.47.1",
    "uvicorn>=0.35.0",
    "vulture>=2.14",
    "xlrd>=2.0.2",
]
